Refactored the new package command generator
Added new Xcode class
Updated tests
