+++ lib/gym/generators/README.md
@@ -1,0 +1,1 @@
+All the classes that generate commands we want to execute
