+++ lib/gym/generators/build_command_generator.rb
