+++ spec/package_command_generator_legacy_spec.rb
@@ -3,16 +3,16 @@ require 'gym/xcodebuild_fixes/watchkit_fix'
 require 'gym/xcodebuild_fixes/package_application_fix'
 
 describe Gym do
-  describe Gym::PackageCommandGeneratorPre7 do
+  describe Gym::PackageCommandGeneratorLegacy, now: true do
     it "works with the example project with no additional parameters" do
       options = { project: "./examples/standard/Example.xcodeproj" }
       Gym.config = FastlaneCore::Configuration.create(Gym::Options.available_options, options)
 
-      result = Gym::PackageCommandGeneratorPre7.generate
+      result = Gym::PackageCommandGeneratorLegacy.generate
       expect(result).to eq([
         "/usr/bin/xcrun /tmp/PackageApplication4Gym -v",
         "''",
-        "-o '#{Gym::PackageCommandGeneratorPre7.ipa_path}'",
+        "-o '#{Gym::PackageCommandGeneratorLegacy.ipa_path}'",
         "exportFormat ipa",
         ""
       ])
@@ -22,13 +22,13 @@ describe Gym do
       options = { project: "./examples/standard/Example.xcodeproj" }
       Gym.config = FastlaneCore::Configuration.create(Gym::Options.available_options, options)
 
-      allow(Gym::PackageCommandGeneratorPre7).to receive(:appfile_path).and_return("Krause's App")
+      allow(Gym::PackageCommandGeneratorLegacy).to receive(:appfile_path).and_return("Krause's App")
 
-      result = Gym::PackageCommandGeneratorPre7.generate
+      result = Gym::PackageCommandGeneratorLegacy.generate
       expect(result).to eq([
         "/usr/bin/xcrun /tmp/PackageApplication4Gym -v",
         "Krause\\'s\\ App",
-        "-o '#{Gym::PackageCommandGeneratorPre7.ipa_path}'",
+        "-o '#{Gym::PackageCommandGeneratorLegacy.ipa_path}'",
         "exportFormat ipa",
         ""
       ])
@@ -43,11 +43,11 @@ describe Gym do
       }
       Gym.config = FastlaneCore::Configuration.create(Gym::Options.available_options, options)
 
-      result = Gym::PackageCommandGenerator.generate
+      result = Gym::PackageCommandGeneratorLegacy.generate
       expect(result).to eq([
         "/usr/bin/xcrun /tmp/PackageApplication4Gym -v",
         "''",
-        "-o '#{Gym::PackageCommandGenerator.ipa_path}'",
+        "-o '#{Gym::PackageCommandGeneratorLegacy.ipa_path}'",
         "exportFormat ipa",
         "--embed './spec/fixtures/dummy.mobileprovision'",
         ""
